Fix error introduced in defaultSorting.
Stupid java naming an int and an <E> method the same name.

+++ src/minecraft/invtweaks/InvTweaksHandlerSorting.java
@@ -388,12 +388,12 @@ public class InvTweaksHandlerSorting extends InvTweaksObfuscation {
                 if (hasToBeMoved(i)) {
                     for (int j = 0; j < size; j++) {
                         if (move(i, j, 1) != -1) {
-                            nextRemaining.remove(j);
+                            nextRemaining.remove((Integer)j);
                             break;
                         }
                     }
                 } else {
-                    nextRemaining.remove(i);
+                    nextRemaining.remove((Integer)i);
                 }
             }
             remaining.clear();
